refactor(lore): address PR #67 review feedback
- Extract DuckDB init + Ollama embedding into a new rag/lore-db.ts module
  (placed alongside lore.ts since rag/ is the documented home for embedding
  stores; state/ is JSON-backed). Both rag/lore.ts and rag/communities.ts
  now import from there. Drops the underscore-prefixed _getLoreDb /
  _openLoreWriteConn / _getLoreEmbedding "private" exports.

- Make seedrandom call defensive: pass { global: false } explicitly to
  guarantee no Math.random replacement, even though the no-options form
  already returns a local PRNG.

- Move graphology-types to devDependencies (types-only, no runtime).

- Cap list_communities limit at 100 (Math.min) to match the schema doc and
  bound GM-tool output regardless of caller request.

- CLAUDE.md: document ANTHROPIC_API_KEY and SCRIBE_SUMMARY_MODEL under the
  scribe env-var list, plus reference the new rag/lore-db.ts and
  rag/communities.ts modules in the supporting-modules list.
