+++ plugins/ironsworn/scribe/package.json
@@ -13,7 +13,6 @@
     "@modelcontextprotocol/sdk": "^1.11.0",
     "graphology": "^0.26.0",
     "graphology-communities-louvain": "^2.0.2",
-    "graphology-types": "^0.24.8",
     "seedrandom": "^3.0.5",
     "yaml": "^2.7.0",
     "zod": "^4.3.6"
@@ -22,6 +21,7 @@
     "typescript": "^5.8.0",
     "@types/node": "^22.0.0",
     "@types/bun": "^1.2.0",
-    "@types/seedrandom": "^3.0.8"
+    "@types/seedrandom": "^3.0.8",
+    "graphology-types": "^0.24.8"
   }
 }
