+++ plugins/ironsworn/scribe/src/rag/lore.ts
@@ -1,12 +1,10 @@
 import { DuckDBInstance } from "@duckdb/node-api";
-import { mkdir } from "node:fs/promises";
-
-// ---------------------------------------------------------------------------
-// Config
-// ---------------------------------------------------------------------------
-
-const OLLAMA_BASE_URL =
-  process.env["OLLAMA_BASE_URL"] ?? "http://localhost:11434";
+import {
+  getLoreDb,
+  openLoreWriteConn,
+  getLoreEmbedding,
+  peekLoreDb,
+} from "./lore-db.js";
 
 // ---------------------------------------------------------------------------
 // Public types
@@ -104,191 +102,6 @@ export function slugify(text: string): string {
 }
 
 // ---------------------------------------------------------------------------
-// Lazy per-campaign DB cache
-// ---------------------------------------------------------------------------
-
-const _dbPromises = new Map<string, Promise<DuckDBInstance>>();
-
-async function initDb(campaignPath: string): Promise<DuckDBInstance> {
-  await mkdir(campaignPath, { recursive: true });
-
-  const instance = await DuckDBInstance.create(`${campaignPath}/lore.duckdb`);
-
-  const conn = await instance.connect();
-  try {
-    await conn.run("INSTALL vss;");
-    await conn.run("LOAD vss;");
-    // HNSW persistence is gated behind an experimental flag in DuckDB. Without
-    // this, HNSW indexes can only be built on in-memory databases — which fails
-    // the moment we try to persist lore to a campaign directory.
-    await conn.run("SET hnsw_enable_experimental_persistence = true;");
-
-    await conn.run(`
-      CREATE TABLE IF NOT EXISTS lore_entities (
-        id         TEXT PRIMARY KEY,
-        canonical  TEXT NOT NULL,
-        aliases    TEXT[] NOT NULL DEFAULT [],
-        type       TEXT NOT NULL,
-        summary    TEXT NOT NULL,
-        content    TEXT NOT NULL DEFAULT '{}',
-        metadata   TEXT NOT NULL DEFAULT '{}',
-        embedding  FLOAT[768] NOT NULL,
-        created_at TEXT NOT NULL,
-        updated_at TEXT NOT NULL
-      )
-    `);
-
-    await conn.run(`
-      CREATE INDEX IF NOT EXISTS lore_embedding_idx
-      ON lore_entities USING HNSW (embedding)
-      WITH (metric = 'cosine')
-    `);
-
-    await conn.run(`
-      CREATE TABLE IF NOT EXISTS lore_relations (
-        from_id    TEXT NOT NULL,
-        to_id      TEXT NOT NULL,
-        relation   TEXT NOT NULL,
-        notes      TEXT,
-        metadata   TEXT NOT NULL DEFAULT '{}',
-        embedding  FLOAT[768],
-        created_at TEXT NOT NULL,
-        PRIMARY KEY (from_id, to_id, relation)
-      )
-    `);
-
-    await conn.run(`
-      CREATE TABLE IF NOT EXISTS lore_provenance (
-        id           TEXT PRIMARY KEY,
-        subject_kind TEXT NOT NULL,
-        subject_id   TEXT NOT NULL,
-        source_kind  TEXT NOT NULL,
-        source_id    TEXT,
-        excerpt      TEXT,
-        confidence   FLOAT,
-        created_at   TEXT NOT NULL
-      )
-    `);
-
-    await conn.run(`
-      CREATE INDEX IF NOT EXISTS lore_provenance_subject_idx
-      ON lore_provenance (subject_kind, subject_id)
-    `);
-
-    // GraphRAG community layer (issue #57). Hierarchical clusters of entities
-    // produced by recompute_communities. member_ids holds direct children:
-    // entity ids at level 0; child community ids at level > 0. embedding is
-    // nullable because summaries are generated only for created/changed
-    // communities — unchanged communities keep their existing summary+embedding
-    // across reruns. No HNSW index is built here; search_lore_global (Phase C)
-    // will add it when global search ships.
-    await conn.run(`
-      CREATE TABLE IF NOT EXISTS lore_communities (
-        id           TEXT PRIMARY KEY,
-        level        INTEGER NOT NULL,
-        parent_id    TEXT,
-        member_ids   TEXT[] NOT NULL DEFAULT [],
-        member_count INTEGER NOT NULL,
-        summary      TEXT NOT NULL DEFAULT '',
-        embedding    FLOAT[768],
-        metadata     TEXT NOT NULL DEFAULT '{}',
-        created_at   TEXT NOT NULL,
-        updated_at   TEXT NOT NULL
-      )
-    `);
-
-    await conn.run(`
-      CREATE INDEX IF NOT EXISTS lore_communities_parent_idx
-      ON lore_communities (parent_id)
-    `);
-
-    await conn.run(`
-      CREATE INDEX IF NOT EXISTS lore_communities_level_idx
-      ON lore_communities (level)
-    `);
-  } finally {
-    conn.closeSync();
-  }
-
-  return instance;
-}
-
-// Exported so the communities module can share the same lazy DB cache without
-// reimplementing the init+migration logic.
-export { getDb as _getLoreDb, openWriteConn as _openLoreWriteConn };
-
-function getDb(campaignPath: string): Promise<DuckDBInstance> {
-  const cached = _dbPromises.get(campaignPath);
-  if (cached !== undefined) return cached;
-
-  const promise = initDb(campaignPath).catch((e) => {
-    _dbPromises.delete(campaignPath);
-    throw e;
-  });
-  _dbPromises.set(campaignPath, promise);
-  return promise;
-}
-
-// DuckDB's hnsw_enable_experimental_persistence flag is connection-scoped, not
-// database-scoped. The initDb connection sets it, but every subsequent write
-// connection opens fresh without it — causing "Duplicate keys not allowed in
-// high-level wrappers" when DuckDB tries to replay buffered HNSW index appends.
-// Every connection that writes to an HNSW-indexed table must set this flag.
-async function openWriteConn(
-  instance: DuckDBInstance,
-): Promise<Awaited<ReturnType<DuckDBInstance["connect"]>>> {
-  const conn = await instance.connect();
-  await conn.run("SET hnsw_enable_experimental_persistence = true;");
-  return conn;
-}
-
-// ---------------------------------------------------------------------------
-// Embeddings
-// ---------------------------------------------------------------------------
-
-export async function _getLoreEmbedding(text: string): Promise<number[]> {
-  return getEmbedding(text);
-}
-
-async function getEmbedding(text: string): Promise<number[]> {
-  let response: Response;
-  try {
-    response = await fetch(`${OLLAMA_BASE_URL}/api/embed`, {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ model: "nomic-embed-text", input: text }),
-    });
-  } catch (e) {
-    const err = e as Error;
-    throw new Error(`Ollama unavailable at ${OLLAMA_BASE_URL}: ${err.message}`);
-  }
-
-  if (!response.ok) {
-    throw new Error(
-      `Ollama embed failed: ${response.status} ${response.statusText}`,
-    );
-  }
-
-  const data = (await response.json()) as { embeddings: number[][] };
-
-  if (!data.embeddings || !Array.isArray(data.embeddings[0])) {
-    throw new Error("Unexpected Ollama response shape");
-  }
-
-  if (data.embeddings[0].length !== 768) {
-    throw new Error(
-      `Expected 768-dim embedding, got ${data.embeddings[0].length}`,
-    );
-  }
-
-  if (!data.embeddings[0].every((v) => typeof v === "number" && isFinite(v))) {
-    throw new Error("Invalid embedding values from Ollama");
-  }
-
-  return data.embeddings[0];
-}
-
-// ---------------------------------------------------------------------------
 // Row mapping helpers
 // ---------------------------------------------------------------------------
 
@@ -370,15 +183,15 @@ export async function upsertLore(
   }
 
   const [embedding, instance] = await Promise.all([
-    getEmbedding(input.summary),
-    getDb(campaignPath),
+    getLoreEmbedding(input.summary),
+    getLoreDb(campaignPath),
   ]);
 
   const embeddingLiteral = `[${embedding.join(",")}]::FLOAT[768]`;
   const now = new Date().toISOString();
   const contentJson = JSON.stringify(input.content ?? {});
 
-  const conn = await openWriteConn(instance);
+  const conn = await openLoreWriteConn(instance);
   try {
     const existingResult = await conn.runAndReadAll(
       `SELECT canonical, aliases, metadata FROM lore_entities WHERE id = ?`,
@@ -482,8 +295,8 @@ export async function searchLore(
   type?: LoreType,
 ): Promise<LoreSearchHit[]> {
   const [embedding, instance] = await Promise.all([
-    getEmbedding(query),
-    getDb(campaignPath),
+    getLoreEmbedding(query),
+    getLoreDb(campaignPath),
   ]);
 
   const embeddingLiteral = `[${embedding.join(",")}]::FLOAT[768]`;
@@ -552,8 +365,8 @@ export async function linkLore(
   campaignPath: string,
   input: LinkLoreInput,
 ): Promise<{ from_id: string; to_id: string; relation: string }> {
-  const instance = await getDb(campaignPath);
-  const conn = await openWriteConn(instance);
+  const instance = await getLoreDb(campaignPath);
+  const conn = await openLoreWriteConn(instance);
   try {
     const fromId = await resolveId(conn, input.from);
     const toId = await resolveId(conn, input.to);
@@ -635,7 +448,7 @@ export async function getLoreGraph(
   const root = await getLore(campaignPath, identifier);
   if (root === null) return null;
 
-  const instance = await getDb(campaignPath);
+  const instance = await getLoreDb(campaignPath);
   const conn = await instance.connect();
   try {
     const visited = new Set<string>([root.id]);
@@ -707,7 +520,7 @@ export async function getLore(
   campaignPath: string,
   identifier: string,
 ): Promise<LoreEntity | null> {
-  const instance = await getDb(campaignPath);
+  const instance = await getLoreDb(campaignPath);
   const needle = identifier.toLowerCase();
 
   const conn = await instance.connect();
@@ -791,7 +604,7 @@ export async function listProvenance(
   subjectKind: "entity" | "relation",
   subjectId: string,
 ): Promise<ProvenanceEntry[]> {
-  const instance = await getDb(campaignPath);
+  const instance = await getLoreDb(campaignPath);
   const conn = await instance.connect();
   try {
     const result = await conn.runAndReadAll(
@@ -851,7 +664,7 @@ export interface LoreRelationExport {
 export async function exportLore(
   campaignPath: string,
 ): Promise<{ entities: LoreEntityExport[]; relations: LoreRelationExport[] }> {
-  const instance = await getDb(campaignPath);
+  const instance = await getLoreDb(campaignPath);
   const conn = await instance.connect();
   try {
     const entRows = (
@@ -911,7 +724,7 @@ export async function exportLore(
  * Safe to call at any time; no-op if the DB has not been opened yet.
  */
 export async function checkpointLore(campaignPath: string): Promise<void> {
-  const cached = _dbPromises.get(campaignPath);
+  const cached = peekLoreDb(campaignPath);
   if (cached === undefined) return;
 
   const instance = await cached;
